improvements on code splitting
