+++ src/screens/Learn/Glossary/Glossary.tsx
@@ -9,7 +9,8 @@ import AppMetaTags from 'components/AppMetaTags/AppMetaTags';
 import { MarkdownContent } from 'components/Markdown';
 import { EventCategory } from 'components/Analytics';
 import PageContent from 'components/PageContent';
-import { glossaryContent, learnPages, Term } from 'cms-content/learn';
+import { glossaryContent, Term } from 'cms-content/learn/glossary';
+import { learnPages } from 'cms-content/learn';
 import Breadcrumbs from 'components/Breadcrumbs';
 import { Anchor } from 'components/TableOfContents';
 import { formatMetatagDate, formatNumericalDate } from 'common/utils';
