+++ src/screens/Learn/Landing/Landing.tsx
@@ -4,11 +4,8 @@ import AppMetaTags from 'components/AppMetaTags/AppMetaTags';
 import { Heading1, Heading2, MarkdownContent } from 'components/Markdown';
 import PageContent from 'components/PageContent';
 import { Anchor } from 'components/TableOfContents';
-import {
-  LandingSection,
-  landingPageContent,
-  learnPages,
-} from 'cms-content/learn';
+import { LandingSection, landingPageContent } from 'cms-content/learn/landing';
+import { learnPages } from 'cms-content/learn';
 import SectionButton, { ButtonTheme } from './SectionButton';
 import { ButtonContainer } from '../Learn.style';
